prepare inventory system; clean up character class
