+++ Necromancy.Server/Chat/Command/Commands/PlayersCommand.cs
@@ -59,7 +59,7 @@ namespace Necromancy.Server.Chat.Command.Commands
                         if (theirCharacter.Name == command[0])
                         {
                             responses.Add(ChatResponse.CommandError(client,
-                                $"{theirCharacter.Name} {theirCharacter.SoulName} is on Map {theirCharacter.MapId} with InstanceID {theirCharacter.InstanceId}"));
+                                $"{theirCharacter.Name} SoulId:{theirCharacter.SoulId} is on Map {theirCharacter.MapId} with InstanceID {theirCharacter.InstanceId}"));
                             soulFound = true;
                         }
                     }
