+++ Necromancy.Server/Model/CharacterModel/CharacterLookup.cs
