+++ Necromancy.Server/Model/ItemModel/Inventory.cs
@@ -1,0 +1,15 @@
+namespace Necromancy.Server.Model.ItemModel
+{
+    public class Inventory
+    {
+        public InventoryItem GetInventoryItem(int inventoryItemId)
+        {
+            return null;
+        }
+
+        public InventoryItem GetInventoryItem(int bagId, short bagSlotIndex)
+        {
+            return null;
+        }
+    }
+}
