+++ Necromancy.Server/Model/NecClient.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics;
 using Arrowgene.Logging;
 using Necromancy.Server.Logging;
+using Necromancy.Server.Model.ItemModel;
 using Necromancy.Server.Packet;
 
 namespace Necromancy.Server.Model
@@ -26,6 +27,7 @@ namespace Necromancy.Server.Model
         public Character Character { get; set; }
         public Channel Channel { get; set; }
         public Map Map { get; set; }
+        public Inventory Inventory { get; set; }
         public Union.Union Union { get; set; }
         public NecConnection AuthConnection { get; set; }
         public NecConnection MsgConnection { get; set; }
