+++ Necromancy.Server/Packet/Area/send_battle_release_attack_pose.cs
@@ -1,6 +1,7 @@
 using Arrowgene.Buffers;
 using Necromancy.Server.Common;
 using Necromancy.Server.Model;
+using Necromancy.Server.Model.CharacterModel;
 using Necromancy.Server.Packet.Id;
 
 namespace Necromancy.Server.Packet.Area
@@ -37,7 +38,7 @@ namespace Necromancy.Server.Packet.Area
                 client);
 
             client.Character.weaponEquipped = false;
-            client.Character.ClearStateBit(0x2);
+            client.Character.ClearStateBit(CharacterState.BattlePose);
         }
     }
 }
