+++ Necromancy.Server/Packet/Receive/RecvItemUpdateEqMask.cs
@@ -8,40 +8,38 @@ namespace Necromancy.Server.Packet.Receive
 {
     public class RecvItemUpdateEqMask : PacketResponse
     {
-        private readonly InventoryItem _invItem;
-        private readonly int _bitmask;
+        private readonly InventoryItem _inventoryItem;
 
-        public RecvItemUpdateEqMask(InventoryItem invItem, int bitmask)
+        public RecvItemUpdateEqMask(InventoryItem inventoryItem)
             : base((ushort) AreaPacketId.recv_item_update_eqmask, ServerType.Area)
         {
-            _invItem = invItem;
-            _bitmask = bitmask;
+            _inventoryItem = inventoryItem;
         }
 
         protected override IBuffer ToBuffer()
         {
             IBuffer res = BufferProvider.Provide();
-    //     res.WriteUInt64(_invItem.InstanceId);
-    //     res.WriteInt32(_bitmask); //Equip bitmask
+            res.WriteUInt64((ulong) _inventoryItem.Id);
+            res.WriteInt32((int) _inventoryItem.Item.EquipmentSlotType);
 
-    //     res.WriteInt32(_invItem.StorageItem.icon);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
+            res.WriteInt32(_inventoryItem.Item.Id);
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
 
-    //     res.WriteInt32(_invItem.StorageItem.icon);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
+            res.WriteInt32(_inventoryItem.Item.Id);
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
 
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0); //bool
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
-    //     res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0); //bool
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
+            res.WriteByte(0);
 
             return res;
         }
